refactor utility functions into separate module
